requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
datasets==2.21.0
dateparser==1.2.0
DateTime==5.5
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docarray==0.40.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
groq==0.10.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.36
langchain-experimental==0.0.64
langchain-groq==0.1.9
langchain-ollama==0.1.2
langchain-text-splitters==0.2.2
langsmith==0.1.104
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
narwhals==1.5.5
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==1.26.4
ollama==0.3.1
openai==1.42.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pinecone==5.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
streamlit==1.38.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.9.4
zope.interface==7.0.3
